refactor(architecture): split type_reachability.py and pr_comment.py (ADR-061)
Batch 1 of the >1500-line abicheck/ file cleanup:

- type_reachability.py (1504 -> 1132 lines): moved
  directly_referenced_stdlib_type_spellings and its three private helpers
  (_alias_confirmed_identities, _stripped_spelling_index,
  _assemble_spellings) into a new sibling, type_reachability_stdlib_spellings.py.
  The public function stays importable from abicheck.type_reachability via
  a dynamic importlib.import_module indirection (the new module needs
  _merged_typedefs/_run_stdlib_reference_scan back from the original, so a
  static re-export would create a type_reachability <-> ..._stdlib_spellings
  cycle -- the same shape serialization.py's own bundle_facts_serialization.py
  split already resolved this way).
- pr_comment.py (1545 -> 882 lines): moved the whole 'Rendering --
  CommentModel -> markdown' half (already marked by the file's own section
  divider) into a new sibling, pr_comment_render.py. render_comment() takes
  an already-built CommentModel and calls nothing in pr_comment.py's own
  parsing half, so pr_comment.py's re-export of render_comment/MARKER/
  DETAIL_LEVELS/GITHUB_COMMENT_LIMIT/_header is a plain static import with
  no cycle. One test (test_pr_comment.py) patched the moved _BODY_BUDGET
  directly on the old module; repointed to the real owner per this repo's
  own D8 migration rule (the test moves with the implementation).

Both new modules registered in architecture/modules.yaml alongside their
siblings. Verified: check_architecture.py 0 errors, mypy clean, ruff clean,
419 targeted tests pass (type_reachability + contract_evaluation stdlib
reachability + full pr_comment/report_integrity suites).
